feat(company): refactor CompanyInformationItem and improve data display (#268)
### TL;DR

Refactored company profile components and introduced a reusable CompanyInformationItem component.

### What changed?

- Created a new `CompanyInformationItem` component in a separate file.
- Updated imports and removed duplicate `CompanyInformationItem` definitions in various files.
- Standardized the display of empty values to show a dash ('-') instead of 'No data'.
- Adjusted the rendering of agent names and dates to use `undefined` for consistency.
- Removed unused imports and variables.

### How to test?

1. Navigate to the company profile pages for different accounts.
2. Verify that all information is displayed correctly, with empty values showing as '-'.
3. Check that dates and agent names are formatted properly.
4. Ensure that the edit mode functionality still works as expected.

### Why make this change?

This refactoring improves code maintainability and consistency across the company profile components. By centralizing the `CompanyInformationItem` component, we reduce code duplication and make it easier to apply future changes uniformly. The standardization of empty value display enhances the user experience by providing a consistent visual representation throughout the interface.
