+++ src/app/(dashboard)/(home)/accounts/(Personnel)/[id]/(company profile)/company-information-item.tsx
@@ -1,0 +1,14 @@
+const CompanyInformationItem = ({
+  label,
+  value,
+}: {
+  label: string
+  value?: string | undefined
+}) => (
+  <div className="flex flex-col py-1">
+    <div className="text-sm font-medium text-muted-foreground">{label}</div>
+    <div className="text-md font-semibold">{value || '-'}</div>
+  </div>
+)
+
+export default CompanyInformationItem
